requirements.txt
discord.py==2.4.0
aiohttp==3.9.5
pytz==2024.1
Pillow==10.4.0